requirements.txt
absl-py==0.15.0
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
clang==5.0
colorama==0.4.5
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==0.10.0
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gensim==4.2.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.0.6
jupyter_core==4.11.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lightgbm==3.3.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
setuptools==65.0.2
six==1.16.0
smart-open==6.0.0
stack-data==0.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==3.7.4.3
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.1
wheel==0.37.1
wrapt==1.12.1
zipp==3.8.1